# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: RALD Ecosystem API — includes Customer Graph (Phase D)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List workspaces the current user belongs to
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Workspace list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "409":
          description: Slug already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace detail
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      responses:
        "200":
          description: Workspace detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "403":
          description: Not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /workspaces/{workspaceId}/members:
    get:
      operationId: listWorkspaceMembers
      tags: [workspaces]
      summary: List workspace members
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceMember"
    post:
      operationId: addWorkspaceMember
      tags: [workspaces]
      summary: Add a member to the workspace
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceMemberInput"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceMember"
        "409":
          description: Already a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /workspaces/{workspaceId}/members/{userId}:
    delete:
      operationId: removeWorkspaceMember
      tags: [workspaces]
      summary: Remove a member from the workspace
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
  /workspaces/{workspaceId}/tags:
    get:
      operationId: listCustomerTags
      tags: [workspaces]
      summary: List all customer tags in the workspace
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      responses:
        "200":
          description: Tag list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerTag"
    post:
      operationId: createCustomerTag
      tags: [workspaces]
      summary: Create a customer tag
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerTagInput"
      responses:
        "201":
          description: Tag created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerTag"
  /workspaces/{workspaceId}/tags/{tagId}:
    delete:
      operationId: deleteCustomerTag
      tags: [workspaces]
      summary: Delete a customer tag
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
        - name: tagId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─────────────────────────────────────────────────────────────────────────────
  # CUSTOMERS  (Phase D — Customer Graph)
  # ─────────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    workspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
      description: Workspace UUID
...
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Workspace:
      type: object
      required: [id, name, slug, plan, isActive, ownerId, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
          enum: [starter, growth, enterprise]
        isActive:
          type: boolean
        ownerId:
          type: string
        memberRole:
          type: string
          enum: [owner, admin, member, viewer]
          description: Role of the requesting user (present when listing user workspaces)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkspaceInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          maxLength: 120
        slug:
          type: string
          maxLength: 60
          pattern: "^[a-z0-9-]+$"
    WorkspaceMember:
      type: object
      required: [id, workspaceId, userId, role, createdAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        userId:
          type: string
        role:
          type: string
          enum: [owner, admin, member, viewer]
        createdAt:
          type: string
          format: date-time
    WorkspaceMemberInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [owner, admin, member, viewer]
          default: member

    # ── Customer schemas ────────────────────────────────────────────────────────
...
    CustomerTag:
      type: object
      required: [id, workspaceId, name, createdAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        color:
          type: ["string", "null"]
          description: Hex color string (e.g. #10b981)
        description:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CustomerTagInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 80
        color:
          type: string
          description: Hex color string
        description:
          type: string
...