# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RALD Ecosystem API — includes Customer Graph (Phase D)
 * OpenAPI spec version: 0.2.0
 */

export interface ApiError {
  error: string;
}